+++ ansible/playbooks/roles/restmail/files/restmail-email.conf
@@ -1,5 +1,5 @@
 [program:restmail-email]
-command=node emailserver.js
+command=node bin/emailserver.js
 directory=/home/app/restmail
 autostart=true
 autorestart=unexpected
@@ -13,3 +13,4 @@ stderr_logfile=/var/log/restmail-email-err.log
 stderr_logfile_maxbytes=100MB
 stderr_logfile_backups=10
 user=app
+environment=TMP_DIR=/home/app/specials,RESTMAIL_RCPT_TO_DOMAIN_WHITELIST="localhost,restmail.com,restmail.net,example.com",RESTMAIL_MAXIMUM_RCPT_TO=5000,RESTMAIL_EXPIRE_AFTER=21600
