+++ ansible/playbooks/roles/restmail/files/restmail-web.conf
@@ -1,5 +1,5 @@
 [program:restmail-web]
-command=node webserver.js
+command=node bin/webserver.js
 directory=/home/app/restmail
 autostart=true
 autorestart=unexpected
